se creo carpeta componentes en el directorio src, se agrega customfrom.
Se migro el formulario desde TallerHome a CustomForm,

correccion de ubicacion de checkboxes

Implementacion de JSX ItemData que devuelve caption en negrita y el resto en letra comun.
