+++ src/shared/interface.ts
@@ -1,0 +1,10 @@
+export interface TallerHomeDTO {
+    title: string;
+    resume: string;
+    location: string;
+    date: string;
+    hour: string;
+    caption: string;
+    contacPhone: string;
+    contact: string;
+}
